fixed c++ 11 issues (for clang/gfortran on macos). use installed HDF5 1.8.11 (unless chombo)
svn path=/vcell/trunk/numerics/; revision=21112


Former-commit-id: 3720431fded55f71bab2df66aec130fd1fc42fcb

+++ MBSolver/Solver/src/CoordVect.cpp
@@ -43,7 +43,7 @@ CoordVect::CoordVect(const std::array<double, DIM>& p)
 
 CoordVect::CoordVect(const moving_boundary::MeshElementNode& me)
 {
-	double worldValues[DIM] = {me(spatial::cX), me(spatial::cY)};
+	double worldValues[DIM] = {(double)me(spatial::cX), (double)me(spatial::cY)};
 	moving_boundary::World<moving_boundary::CoordinateType, 2>::get().toProblemDomain(worldValues, vect);
 }
